refactor(slack-bot): extract slack-api.js from index.js (behavior-preserving)
